requirements.txt
tensorflow
numpy
pillow
flask
gunicorn
